starred_messages: Extract starred_messages_ui.js.
