+++ web/src/starred_messages.js
@@ -1,8 +1,5 @@
 import * as message_store from "./message_store";
 import {page_params} from "./page_params";
-import * as popover_menus from "./popover_menus";
-import * as top_left_corner from "./top_left_corner";
-import {user_settings} from "./user_settings";
 
 export const starred_ids = new Set();
 
@@ -12,24 +9,18 @@ export function initialize() {
     for (const id of page_params.starred_messages) {
         starred_ids.add(id);
     }
-
-    rerender_ui();
 }
 
 export function add(ids) {
     for (const id of ids) {
         starred_ids.add(id);
     }
-
-    rerender_ui();
 }
 
 export function remove(ids) {
     for (const id of ids) {
         starred_ids.delete(id);
     }
-
-    rerender_ui();
 }
 
 export function get_count() {
@@ -70,17 +61,4 @@ export function get_count_in_topic(stream_id, topic) {
     });
 
     return messages.length;
-}
-
-export function rerender_ui() {
-    let count = get_count();
-
-    if (!user_settings.starred_message_counts) {
-        // This essentially hides the count
-        count = 0;
-    }
-
-    popover_menus.get_topic_menu_popover()?.hide();
-    popover_menus.get_starred_messages_popover()?.hide();
-    top_left_corner.update_starred_count(count);
 }
