+++ web/tests/message_flags.test.js
@@ -12,11 +12,16 @@ set_global("document", {hasFocus: () => true});
 
 mock_esm("../src/starred_messages", {
     add() {},
+    get_count: () => 5,
     get_starred_msg_ids: () => [1, 2, 3, 4, 5],
     remove() {},
 });
+mock_esm("../src/top_left_corner", {
+    update_starred_count() {},
+});
 
 const message_flags = zrequire("message_flags");
+const starred_messages_ui = zrequire("starred_messages_ui");
 
 run_test("starred", ({override}) => {
     const message = {
@@ -35,7 +40,7 @@ run_test("starred", ({override}) => {
         posted_data = opts.data;
     });
 
-    message_flags.toggle_starred_and_update_server(message);
+    starred_messages_ui.toggle_starred_and_update_server(message);
 
     assert.ok(ui_updated);
 
@@ -52,7 +57,7 @@ run_test("starred", ({override}) => {
 
     ui_updated = false;
 
-    message_flags.toggle_starred_and_update_server(message);
+    starred_messages_ui.toggle_starred_and_update_server(message);
 
     assert.ok(ui_updated);
 
@@ -76,7 +81,7 @@ run_test("starring local echo", () => {
         locally_echoed: true,
     };
 
-    message_flags.toggle_starred_and_update_server(locally_echoed_message);
+    starred_messages_ui.toggle_starred_and_update_server(locally_echoed_message);
 
     // message_live_update.update_starred_view not called
